Splitted target_painter.dart into TargetBuilder.dart and Target.dart.
