+++ hbp_with_firebase_mlkit/lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/material.dart';
 import 'package:camera/camera.dart';
-import 'target_painter.dart';
+import 'TargetBuilder.dart';
 import 'face_painter.dart';
 import 'pointer.dart';
 import 'utils.dart';
